pyproject.toml
[tool.poetry]
name = "kyrsovoi-project"
version = "0.1.0"
description = ""
authors = ["Angelina, Elmira"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
SQLAlchemy = "^2.0.29"
asyncpg = "^0.29.0"
aiogram = "2.25.1"
requests = "^2.31.0"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
